Add rad resource-provider create
This change adds a new command that can register a resource provider, types, and api versions.

Users author a YAML file (manifest), and the `rad` CLI will turn that into a series of API calls.

See: https://github.com/radius-project/design-notes/pull/74 for the somewhat in-progress design of the manifest. It's likely that the manifest will continue to evolve, and we'll update the code in main to match.

This PR contains:

- The new command `rad resource-provider create`
- The parsing and validation logic for the manifest
- Plumbing for calling the resource type APIs from the CLI

Signed-off-by: Ryan Nowak <nowakra@gmail.com>