Add default controller to async controller registry
This change adds a "default" controller capability to the controller registry used by the async worker. This is needed for the dynamic-rp. Because the dynamic-rp is extensible, it processes many different types of operations, it's not possible for us to statically register all of these. Instead, the dynamic-rp will use this new default capability to register a single controller with dynamic functionality.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>